重大变更，使用SugarTable

+++ Yuebon.NetCore/Yuebon.Commons/Page/PageResult.cs
@@ -20,14 +20,14 @@ namespace Yuebon.Commons.Pages
             this.ErrMsg = msg;
             this.ResData =rows ;
         }
-        public PageResult(long currentPage, long totalItems, long itemsPerPage)
+        public PageResult(int currentPage, int totalItems, int itemsPerPage)
         {
             CurrentPage = currentPage;
             TotalItems = totalItems;
             ItemsPerPage = itemsPerPage;
         }
 
-        public PageResult(long currentPage, long totalPages, long totalItems, long itemsPerPage, List<T> items, object context) : this(currentPage, totalPages, totalItems)
+        public PageResult(int currentPage, int totalPages, int totalItems, int itemsPerPage, List<T> items, object context) : this(currentPage, totalPages, totalItems)
         {
             ItemsPerPage = itemsPerPage;
             Items = items;
@@ -37,22 +37,22 @@ namespace Yuebon.Commons.Pages
         /// <summary>
         /// 当前页码。
         /// </summary>
-        public long CurrentPage { get; set; }
+        public int CurrentPage { get; set; }
 
         /// <summary>
         /// 总页码数。
         /// </summary>
-        public long TotalPages { get; set; }
+        public int TotalPages { get; set; }
 
         /// <summary>
         /// 记录总数。
         /// </summary>
-        public long TotalItems { get; set; }
+        public int TotalItems { get; set; }
 
         /// <summary>
         /// 每页数量。
         /// </summary>
-        public long ItemsPerPage { get; set; }
+        public int ItemsPerPage { get; set; }
 
         /// <summary>
         /// 当前结果集。
